🐛 Bug : 네비게이터 수정
