+++ app/(main)/exchange-diary/exchange-diary-tab.tsx
@@ -11,10 +11,21 @@ import React from "react"
 export const ExchangeDiaryTab = () => {
   const { push } = useRouter()
   const sendPush = useSendPush()
+  const onClick = () => {
+    const token = localStorage.getItem("token")
+    sendPush({
+      title: "누군가로 부터 새로운 일기가 왔어요",
+      message: "PODA에 들어와서 확인해보세요",
+      link: "https://poda.vercel.app",
+      token: token as string,
+    })
+  }
 
   return (
     <div className="w-full h-full">
-      <div className="pt-[150px] flex flex-col justify-between items-center space-y-3">
+      <div
+        onClick={onClick}
+        className="pt-[150px] flex flex-col justify-between items-center space-y-3">
         <Image src="/assets/truck.gif" width={300} height={300} alt="truck" />
         <h1 className="text-2xl font-black leading-7 text-primary text-center">
           당신의 일기를
